+++ src/modules/auth/forms/forgot-password.tsx
@@ -4,9 +4,8 @@ import Link from 'next/link'
 import { TFunction, useTranslation } from 'next-i18next'
 import * as yup from 'yup'
 
+import { StyledButton, StyledForm, StyledLink } from '@/components/forms/styles'
 import { ForgotPasswordDto } from '@/modules/auth/auth'
-
-import { StyledButton, StyledForm, StyledLink } from './styles'
 
 function generateValidationSchema(
   t: TFunction,
