+++ src/modules/clubs/forms/fields.tsx
@@ -3,12 +3,11 @@ import { Field, useFormikContext } from 'formik'
 import { useTranslation } from 'next-i18next'
 import React from 'react'
 
+import { CountriesCombo } from '@/components/selects/countries-combo'
+import { RegionsCombo } from '@/components/selects/regions-combo'
 import { CreateClubDto, UpdateClubDto } from '@/modules/clubs/types'
 import { CountryDto } from '@/modules/countries/types'
 import { RegionDto } from '@/modules/regions/types'
-
-import { CountriesCombo } from '../../selects/countries-combo'
-import { RegionsCombo } from '../../selects/regions-combo'
 
 interface IFieldsProps {
   regionsData: RegionDto[]
