+++ src/modules/players/details-card.tsx
