+++ src/modules/teams/details-card.tsx
