+++ src/modules/teams/forms/filter.tsx
@@ -1,52 +1,49 @@
-import { Grid, TextField } from '@mui/material'
+import { TextField } from '@mui/material'
 import { styled } from '@mui/material/styles'
 import { Field, Form, Formik } from 'formik'
 import { CheckboxWithLabel } from 'formik-mui'
 import { useTranslation } from 'next-i18next'
 
+import { Container } from '@/components/forms/container'
+import { FilterFormActions } from '@/components/forms/filter-form-actions'
+import { ClubsCombo } from '@/components/selects/clubs-combo'
 import { CompetitionGroupsCombo } from '@/components/selects/competition-groups-combo'
 import { CompetitionsCombo } from '@/components/selects/competitions-combo'
 import { CountriesCombo } from '@/components/selects/countries-combo'
-import { FootedSelect } from '@/components/selects/footed-select'
-import { PlayersPositionCombo } from '@/components/selects/player-positions-combo'
-import { TeamsCombo } from '@/components/selects/teams-combo'
+import { RegionsCombo } from '@/components/selects/regions-combo'
+import { ClubBasicDataDto, ClubsFiltersDto } from '@/modules/clubs/types'
 import { CompetitionGroupBasicDataDto } from '@/modules/competition-groups/types'
 import { CompetitionBasicDataDto } from '@/modules/competitions/types'
 import { CountryDto } from '@/modules/countries/types'
-import { PlayerPositionDto } from '@/modules/player-positions/types'
-import { PlayersFiltersDto } from '@/modules/players/types'
-import { TeamBasicDataDto } from '@/modules/teams/types'
-
-import { Container } from '../container'
-import { FilterFormActions } from '../filter-form-actions'
+import { RegionDto } from '@/modules/regions/types'
 
 const StyledCheckboxContainer = styled('div')(() => ({
   display: 'flex',
   justifyContent: 'center',
 }))
 
-interface IPlayersFilterFormProps {
+type ITeamsFilterFormProps = {
+  regionsData: RegionDto[]
   countriesData: CountryDto[]
-  positionsData: PlayerPositionDto[]
-  teamsData: TeamBasicDataDto[]
+  clubsData: ClubBasicDataDto[]
   competitionsData: CompetitionBasicDataDto[]
   competitionGroupsData: CompetitionGroupBasicDataDto[]
-  filters: PlayersFiltersDto
-  onFilter: (data: PlayersFiltersDto) => void
+  filters: ClubsFiltersDto
+  onFilter: (data: ClubsFiltersDto) => void
   onClearFilters: () => void
 }
 
-export const PlayersFilterForm = ({
+export const TeamsFilterForm = ({
+  regionsData,
   countriesData,
-  teamsData,
-  positionsData,
+  clubsData,
   competitionsData,
   competitionGroupsData,
   filters,
   onFilter,
   onClearFilters,
-}: IPlayersFilterFormProps) => {
-  const { t } = useTranslation(['common', 'players'])
+}: ITeamsFilterFormProps) => {
+  const { t } = useTranslation()
 
   return (
     <Formik
@@ -62,53 +59,26 @@ export const PlayersFilterForm = ({
               as={TextField}
               variant="outlined"
               fullWidth
-              label={t('NAME_OR_SURNAME')}
+              label={t('NAME')}
               size="small"
             />
-            <Grid container spacing={2}>
-              <Grid item xs={6}>
-                <Field
-                  name="bornAfter"
-                  as={TextField}
-                  type="number"
-                  variant="outlined"
-                  fullWidth
-                  label={t('BORN_AFTER')}
-                  size="small"
-                  inputProps={{ min: 1980, max: 2020 }}
-                />
-              </Grid>
-              <Grid item xs={6}>
-                <Field
-                  name="bornBefore"
-                  as={TextField}
-                  type="number"
-                  variant="outlined"
-                  fullWidth
-                  label={t('BORN_BEFORE')}
-                  size="small"
-                  inputProps={{ min: 1980, max: 2020 }}
-                />
-              </Grid>
-            </Grid>
-            <FootedSelect name="footed" label={t('FOOTED')} size="small" />
             <CountriesCombo
               name="countryIds"
               data={countriesData}
               label={t('COUNTRIES')}
               multiple
             />
-            <PlayersPositionCombo
-              name="positionIds"
-              data={positionsData}
-              label={t('POSITIONS')}
+            <RegionsCombo
+              name="regionIds"
+              data={regionsData}
+              label={t('REGIONS')}
               multiple
             />
-            <TeamsCombo
-              data={teamsData}
-              name="teamIds"
-              label={t('TEAM')}
-              multiple
+            <ClubsCombo
+              data={clubsData}
+              name="clubId"
+              label={t('CLUB')}
+              size="small"
             />
             <CompetitionsCombo
               name="competitionIds"
@@ -127,7 +97,7 @@ export const PlayersFilterForm = ({
                 component={CheckboxWithLabel}
                 type="checkbox"
                 name="isLiked"
-                Label={{ label: t('players:LIKED_ONLY') }}
+                Label={{ label: 'Tylko polubione' }}
               />
             </StyledCheckboxContainer>
             <FilterFormActions handleClearFilter={onClearFilters} />
