+++ src/pages/clubs/edit/[slug].tsx
@@ -2,10 +2,10 @@ import { useTranslation } from 'next-i18next'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 
 import { ErrorContent } from '@/components/error/error-content'
-import { EditClubForm } from '@/components/forms/club/edit-club'
 import { Loader } from '@/components/loader/loader'
 import { PageHeading } from '@/components/page-heading/page-heading'
 import { withSessionSsr } from '@/modules/auth/session'
+import { EditClubForm } from '@/modules/clubs/forms/edit'
 import { useUpdateClub } from '@/modules/clubs/hooks'
 import { ClubDto } from '@/modules/clubs/types'
 import { useCountriesList } from '@/modules/countries/hooks'
