+++ src/pages/clubs/index.tsx
@@ -4,13 +4,13 @@ import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import { useState } from 'react'
 
 import { Fab } from '@/components/fab/fab'
-import { ClubsFilterForm } from '@/components/forms/club/clubs-filter-form'
 import { Loader } from '@/components/loader/loader'
 import { ConfirmationModal } from '@/components/modals/confirmation-modal'
 import { PageHeading } from '@/components/page-heading/page-heading'
 import { ClubsTable } from '@/components/tables/clubs'
 import { ClubsTableRow } from '@/components/tables/rows/clubs-row'
 import { withSessionSsr } from '@/modules/auth/session'
+import { ClubsFilterForm } from '@/modules/clubs/forms/filter'
 import { useClubs, useDeleteClub } from '@/modules/clubs/hooks'
 import { ClubsFiltersDto, ClubsSortBy } from '@/modules/clubs/types'
 import { useCountriesList } from '@/modules/countries/hooks'
