+++ src/pages/players/[slug].tsx
@@ -2,12 +2,12 @@ import { Typography } from '@mui/material'
 import { useTranslation } from 'next-i18next'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 
-import { PlayerDetialsCard } from '@/components/details-cards/player'
 import { ErrorContent } from '@/components/error/error-content'
 import { PageHeading } from '@/components/page-heading/page-heading'
 import { TeamAffiliationsTableRow } from '@/components/tables/rows/team-affiliations-row'
 import { TeamAffiliationsTable } from '@/components/tables/team-affiliations'
 import { withSessionSsr } from '@/modules/auth/session'
+import { PlayerDetialsCard } from '@/modules/players/details-card'
 import { PlayerDto } from '@/modules/players/types'
 import { useTeamAffiliations } from '@/modules/team-affiliations.ts/hooks'
 import { TeamAffiliationsSortBy } from '@/modules/team-affiliations.ts/types'
