+++ src/pages/players/edit/[slug].tsx
@@ -2,12 +2,12 @@ import { useTranslation } from 'next-i18next'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 
 import { ErrorContent } from '@/components/error/error-content'
-import { EditPlayerForm } from '@/components/forms/player/edit-player'
 import { Loader } from '@/components/loader/loader'
 import { PageHeading } from '@/components/page-heading/page-heading'
 import { withSessionSsr } from '@/modules/auth/session'
 import { useCountriesList } from '@/modules/countries/hooks'
 import { usePlayerPositionsList } from '@/modules/player-positions/hooks'
+import { EditPlayerForm } from '@/modules/players/forms/edit'
 import { useUpdatePlayer } from '@/modules/players/hooks'
 import { PlayerDto } from '@/modules/players/types'
 import { useTeamsList } from '@/modules/teams/hooks'
