+++ src/pages/teams/[slug].tsx
@@ -2,7 +2,6 @@ import { Typography } from '@mui/material'
 import { useTranslation } from 'next-i18next'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 
-import { TeamDetailsCard } from '@/components/details-cards/team'
 import { ErrorContent } from '@/components/error/error-content'
 import { PageHeading } from '@/components/page-heading/page-heading'
 import { CompetitionParticipationsTable } from '@/components/tables/competition-participations'
@@ -10,6 +9,7 @@ import { CompetitionParticipationsTableRow } from '@/components/tables/rows/comp
 import { withSessionSsr } from '@/modules/auth/session'
 import { useCompetitionParticipations } from '@/modules/competition-participations/hooks'
 import { CompetitionParticipationsSortBy } from '@/modules/competition-participations/types'
+import { TeamDetailsCard } from '@/modules/teams/details-card'
 import { TeamDto } from '@/modules/teams/types'
 import { getTeamBySlug } from '@/services/api/methods/teams'
 import { ApiError } from '@/types/common'
