+++ src/pages/teams/edit/[slug].tsx
@@ -2,11 +2,11 @@ import { useTranslation } from 'next-i18next'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 
 import { ErrorContent } from '@/components/error/error-content'
-import { EditTeamForm } from '@/components/forms/team/edit-team'
 import { Loader } from '@/components/loader/loader'
 import { PageHeading } from '@/components/page-heading/page-heading'
 import { withSessionSsr } from '@/modules/auth/session'
 import { useClubsList } from '@/modules/clubs/hooks'
+import { EditTeamForm } from '@/modules/teams/forms/edit'
 import { useUpdateTeam } from '@/modules/teams/hooks'
 import { TeamDto } from '@/modules/teams/types'
 import { getTeamBySlug } from '@/services/api/methods/teams'
